requirements.txt
fastapi>=0.115
uvicorn>=0.34
python-multipart>=0.0.18
streamlit>=1.42
openai>=1.68
langchain>=0.3
langchain-text-splitters>=0.3
langgraph>=0.3
PyMuPDF>=1.25
pdfplumber>=0.11
chromadb>=0.6
python-dotenv>=1.1
httpx>=0.28
